task1/requirements.txt
pandas
scikit-learn
google-generativeai
python-dotenv
tqdm
ipykernel
nbformat
